# packages/tesslate-marketplace/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Tesslate Federated Marketplace `/v1`
  version: v1
  description: |
    Federated marketplace wire protocol for OpenSail / Tesslate Studio.

    Every response includes `X-Tesslate-Hub-Id` and
    `X-Tesslate-Hub-Api-Version: v1` headers. Endpoints whose capability
    isn't advertised in `/v1/manifest.capabilities[]` return a typed
    `501 unsupported_capability` JSON envelope.
  contact:
    name: Tesslate
    email: marketplace@tesslate.com
servers:
  - url: http://localhost:8800
    description: Local dev
  - url: https://marketplace.tesslate.com
    description: Tesslate Official
...
paths:
...
  /v1/publish/{kind}:
    post:
      tags: [publish]
      security: [{ bearer: [] }]
      parameters:
        - $ref: "#/components/parameters/Kind"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "201":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /v1/publish/{kind}/{slug}/versions/{version}:
    post:
      tags: [publish]
      security: [{ bearer: [] }]
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
        - $ref: "#/components/parameters/Version"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "201":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /v1/submissions/{id}:
    get:
      tags: [publish]
      security: [{ bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /v1/submissions/{id}/withdraw:
    post:
      tags: [publish]
      security: [{ bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    Kind:
      in: path
      name: kind
      required: true
      schema:
        type: string
        enum: [agent, skill, mcp_server, base, app, theme, workflow_template]
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string }
    Version:
      in: path
      name: version
      required: true
      schema: { type: string }
...
  schemas:
    Pricing:
      type: object
      properties:
        pricing_type:
          type: string
          enum: [free, paid, subscription]
        price_cents: { type: integer }
        currency: { type: string }
        stripe_price_id: { type: string, nullable: true }
        interval: { type: string, nullable: true }
        extras: { type: object, additionalProperties: true, nullable: true }
...
    PublishItem:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        long_description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
        features: { type: array, items: { type: string } }
        tech_stack: { type: array, items: { type: string } }
        extra_metadata: { type: object, additionalProperties: true }
        creator_handle: { type: string, nullable: true }
        git_repo_url: { type: string, nullable: true }
        homepage_url: { type: string, nullable: true }
        pricing: { $ref: "#/components/schemas/Pricing" }
    PublishVersion:
      type: object
      required: [version]
      properties:
        version: { type: string }
        changelog: { type: string, nullable: true }
        manifest: { type: object, additionalProperties: true, nullable: true }
        bundle_b64: { type: string, nullable: true }
        pricing: { $ref: "#/components/schemas/Pricing" }
    PublishRequest:
      type: object
      required: [item, version]
      properties:
        item: { $ref: "#/components/schemas/PublishItem" }
        version: { $ref: "#/components/schemas/PublishVersion" }
    SubmissionCheck:
      type: object
      properties:
        stage: { type: string }
        name: { type: string }
        status:
          type: string
          enum: [passed, failed, warning, errored]
        message: { type: string, nullable: true }
        details: { type: object, additionalProperties: true, nullable: true }
        created_at: { type: string, format: date-time }
    Submission:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        state: { type: string }
        stage: { type: string }
        decision: { type: string, nullable: true }
        decision_reason: { type: string, nullable: true }
        submitter_handle: { type: string, nullable: true }
        bundle_sha256: { type: string, nullable: true }
        bundle_size_bytes: { type: integer, nullable: true }
        item_id: { type: string, nullable: true }
        item_version_id: { type: string, nullable: true }
        checks:
          type: array
          items: { $ref: "#/components/schemas/SubmissionCheck" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...